+++ tests/test_migrate.py
@@ -1,0 +1,5 @@
+from tortoise.contrib import test
+
+
+class TestMigrate(test.TruncationTestCase):
+    pass