+++ tests/test_utils.py
@@ -1,6 +1,0 @@
-from unittest import TestCase
-
-
-class TestUtils(TestCase):
-    def test_get_app_connection(self):
-        pass